requirements.txt
```pip-requirements
torch==2.9.1
torchvision==0.24.1
torchaudio==2.9.1
opencv-python==4.12.0.88
numpy==2.2.6
matplotlib==3.10.7
pillow==12.0.0
tqdm==4.67.1
pyyaml==6.0.3
requests==2.32.5
```
